feature: Implement object-store-based ID allocator for globally and monotonically. (#63)
to #65 

Developed an ID allocator in SmartEngine to ensure globally unique and
monotonically increasing IDs for a serias of data and metadata objects.
The allocator leverages object storage's condition write capabilities to
guarantee atomic ID allocation. Supports specifying a range of IDs to be
allocated in a single operation, reducing interactions with the object
storage. Solves the challenge of maintaining ID uniqueness and order in
scenarios involving switch over, fail over, and multiple instances
sharing the same data.