enhancement: make the failure message of creating directory clearer. (#71)
It needs to create several directories during the startup process of
SmartEngine, such as the data directory, the WAL directory, and the
persistent cache directory, etc. The directory creating operation may
fail for several reasons, like permission denied, space exhausted. We
should print the detailed reason for the failure.
to #66